fix: migration 081 — convert cost columns REAL → NUMERIC(20,8)
db-005 findings from 18.4.2 database audit. Converts estimated_cost in
token_usage_records and cost_amount in ai_usage_operational_records from
REAL (IEEE 754 float) to NUMERIC(20,8) for exact decimal accounting.
Uses table-reconstruction pattern with PRAGMA legacy_alter_table to
sidestep SQLite 3.26+ view-validation on RENAME; full rationale in
migration file. db-005-suppress comments added to earlier migrations
that originally declared these columns as REAL.

+++ core/event_store/migrations/042_token_usage_source_refs.sql
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS token_usage_records (
     output_tokens INTEGER NOT NULL DEFAULT 0,
     cached_tokens INTEGER NOT NULL DEFAULT 0,
     total_tokens INTEGER NOT NULL DEFAULT 0,
-    estimated_cost REAL NOT NULL DEFAULT 0,
+    estimated_cost NUMERIC(20, 8) NOT NULL DEFAULT 0,
     purpose TEXT,
     created_at TEXT NOT NULL DEFAULT (datetime('now')),
     source_refs_json TEXT NOT NULL DEFAULT '[]',
